pyproject.toml
[tool.poetry]
name = "django-upload"
version = "0.1.0"
description = ""
authors = ["jasonxie <x5758x@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
extend-exclude = '''
/(
  | migrations
)/
'''
